* fixed warnings

+++ src/ces/IEntity.cpp
@@ -114,7 +114,7 @@ bool IEntity::HasComponent(const std::type_index& id)
 uint32_t IEntity::ComponentsNum(const std::type_index& id)
 {
     TComponentsList& components = m_Components[id];
-    return components.size();
+    return static_cast<uint32_t>(components.size());
 }
 
 void IEntity::AddChild(IEntityPtr entity)
